backend/requirements.txt
flask>=3.0.0
google-cloud-firestore>=2.13.0
pydantic>=2.6.0
pytest>=8.0.0
python-dotenv>=1.0.1
